fix(core): widen led2Pin type to u16 and add AI agent CLI guide
The device firmware uses 65535 as a sentinel value for led2Pin (meaning
"unset"), which overflows u8. This caused config backup to fail with a
deserialization error on real devices. Changed to u16 to match the
actual device protocol.

Also adds CLAUDE.md with a comprehensive CLI usage guide for AI agents,
documenting correct parameter names, raw protocol commands, common
pitfalls, and OTA procedures based on real device testing.

+++ crates/rtls-link-core/src/types.rs
@@ -251,9 +251,9 @@ pub struct AnchorConfig {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct AppConfig {
-    /// LED 2 GPIO pin
+    /// LED 2 GPIO pin (65535 = unset)
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub led2_pin: Option<u8>,
+    pub led2_pin: Option<u16>,
     /// LED 2 state (0 or 1)
     #[serde(skip_serializing_if = "Option::is_none")]
     pub led2_state: Option<u8>,
